requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
requests>=2.32.0
python-dotenv>=1.0.0
click>=8.1.0
